Modulize router
