+++ services/meta-data-repository/src/route/domains/delete.js
@@ -1,0 +1,29 @@
+const express = require('express');
+const logger = require('@basaas/node-logger');
+const { domainOwnerOrAllowed } = require('../../middleware/permission');
+const conf = require('../../conf');
+const { DomainDAO } = require('../../dao');
+const {
+    transformDbResults,
+} = require('../../transform');
+
+const log = logger.getLogger(`${conf.logging.namespace}/domains:delete`);
+
+const router = express.Router();
+
+router.delete('/:id', domainOwnerOrAllowed({
+    permissions: ['not.defined'],
+}), async (req, res, next) => {
+    try {
+        res.send({
+            data: transformDbResults(await DomainDAO.delete(req.params.id)),
+        });
+    } catch (err) {
+        log.error(err);
+        next({
+            status: 500,
+        });
+    }
+});
+
+module.exports = router;
